Make companion9x compile with recent changes

+++ open9x.cpp
@@ -647,7 +647,7 @@ void incRotaryEncoder(uint8_t idx, int8_t inc)
 {
   g_rotenc[idx] += inc;
   int16_t *value = &(phaseaddress(getRotaryEncoderFlightPhase(idx))->rotaryEncoders[idx]);
-  *value = limit((int16_t)-1024, *value + (inc * 8), (int16_t)+1024);
+  *value = limit((int16_t)-1024, (int16_t)(*value + (inc * 8)), (int16_t)+1024);
   eeDirty(EE_MODEL);
 }
